+++ src/main/kotlin/GRBundle.kt
@@ -1,0 +1,8 @@
+import com.intellij.AbstractBundle
+import org.jetbrains.annotations.PropertyKey
+
+private const val BUNDLE = "GoldenRatioBundle"
+
+object GRBundle : AbstractBundle(BUNDLE) {
+    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) = getMessage(key, *params)
+}