+++ src/main/kotlin/actions/GoldenRatioApplyAction.kt
@@ -5,12 +5,13 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.fileEditor.impl.FileEditorManagerImpl
 import com.intellij.openapi.project.DumbAwareAction
+import GRBundle.message
 
-class GoldenRatioApplyAction : DumbAwareAction() {
+class GoldenRatioApplyAction : DumbAwareAction(message("actions.applyText")) {
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return
         val manager = FileEditorManager.getInstance(project) as? FileEditorManagerImpl ?: return
         val editor = manager.selectedEditor ?: return
-        GoldenRatioPlugin.applyGoldenRatio(manager, editor)
+        GoldenRatioPlugin.INSTANCE.applyRatio(manager, editor)
     }
 }