+++ src/main/kotlin/actions/GoldenRatioToggleAction.kt
@@ -5,18 +5,19 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.fileEditor.impl.FileEditorManagerImpl
 import com.intellij.openapi.project.DumbAwareToggleAction
+import GRBundle.message
 
-class GoldenRatioToggleAction : DumbAwareToggleAction() {
-    override fun isSelected(e: AnActionEvent): Boolean = GoldenRatioPlugin.isAutoEnabled
+class GoldenRatioToggleAction : DumbAwareToggleAction(message("actions.toggleText")) {
+    override fun isSelected(e: AnActionEvent): Boolean = GoldenRatioPlugin.INSTANCE.autoEnabled
 
     override fun setSelected(e: AnActionEvent, state: Boolean) {
-        GoldenRatioPlugin.isAutoEnabled = state
+        GoldenRatioPlugin.INSTANCE.autoEnabled = state
 
         if (state) {
             val project = e.project ?: return
             val manager = FileEditorManager.getInstance(project) as? FileEditorManagerImpl ?: return
             val editor = manager.selectedEditor ?: return
-            GoldenRatioPlugin.applyGoldenRatio(manager, editor)
+            GoldenRatioPlugin.INSTANCE.applyRatio(manager, editor)
         }
     }
 }