+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - ability to toggle auto GoldenRation on and off
 - ability to manually invoke GoldenRatio to resize the current pane even if disabled
+- settings page that allows the enabled status and target ratio to be configured
 
 ## [1.0.0] - 2022-06-17
 ### Added